refactor(redux): migrate error handling
