+++ src/reducers/dialogs/index.js
@@ -9,6 +9,7 @@ import appSettingsReducer from '~/features/settings/dialog';
 import authenticationReducer from '~/features/servers/authentication-dialog';
 import deauthenticationReducer from '~/features/servers/deauthentication-dialog';
 import dockDetailsDialogReducer from '~/features/docks/details';
+import errorHandlingReducer from '~/features/error-handling/slice';
 import featureEditorReducer from '~/features/map-features/editor';
 import geofenceSettingsReducer from '~/features/geofence/slice';
 import layerSettingsReducer from '~/features/map/layer-settings-dialog';
@@ -16,8 +17,6 @@ import promptReducer from '~/features/prompt/slice';
 import savedLocationEditorReducer from '~/features/saved-locations/editor';
 import serverSettingsReducer from '~/features/servers/server-settings-dialog';
 import uavDetailsDialogReducer from '~/features/uavs/details';
-
-import errorHandlingReducer from './error-handling';
 
 /**
  * The reducer function that is responsible for handling all dialog-related
